+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2021-02-21
+
+### Added
+
+-   combination of POC and dataCollector to index, works against live charts
+-   Account and CoinData class objects for tracking live state
+
+### Updated
+
+-   POC and other helper files
+
 ## [0.3.1] - 2021-02-20
 
 ### Updated
@@ -79,6 +90,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Working demo and API connections
 
+[0.4.0]: https://github.com/JesseChamberlain/binance-trader/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/JesseChamberlain/binance-trader/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/JesseChamberlain/binance-trader/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/JesseChamberlain/binance-trader/compare/v0.1.5...v0.2.0